requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
jinja2>=3.1.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
alembic>=1.13.0
